+++ internal/app/handlers/handler_initializer_test.go
@@ -25,5 +25,26 @@ func TestInitializeUserHandlers(t *testing.T) {
 	}
 
 	mock.ExpectClose()
+}
 
+func TestInitializeURLHandlers(t *testing.T) {
+	db, mock, err := sqlmock.New()
+
+	if err != nil {
+		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
+	}
+
+	defer db.Close()
+
+	urlHandler, err := InitializeURLHandlers(db)
+
+	if err != nil {
+		t.Errorf("Error initializing URL handlers: %s", err)
+	}
+
+	if urlHandler == nil {
+		t.Errorf("URL handler is nil")
+	}
+
+	mock.ExpectClose()
 }