+++ internal/app/models/user/user.go
@@ -1,4 +1,4 @@
-package models
+package user_model
 
 import (
 	"errors"
@@ -8,7 +8,7 @@ import (
 var ErrUserNotFound = errors.New("auth not found")
 var ErrUserAlreadyExists = errors.New("auth already exists")
 
-// User represents a auth entity in the application.
+// User represents an auth entity in the application.
 type User struct {
 	ID               uint      `json:"id"`
 	Username         string    `json:"username"`