+++ internal/app/services/auth/auth_service_test.go
@@ -2,7 +2,7 @@ package auth_service
 
 import (
 	"testing"
-	"url-shortener/internal/app/models"
+	"url-shortener/internal/app/models/user"
 	"url-shortener/internal/mocks"
 
 	"github.com/stretchr/testify/assert"
@@ -16,23 +16,29 @@ func TestCreateUser(t *testing.T) {
 
 	t.Run("Create User Successfully", func(t *testing.T) {
 		// Call the CreateUser method
-		token, err := userService.CreateUser("testuser", "password123")
+		err := userService.CreateUser("testuser", "password123")
 
 		// Assertions
 		assert.NoError(t, err)
-		assert.NotEmpty(t, token)
 
 	})
 
 	t.Run("Failed to Create User", func(t *testing.T) {
 		// Call the CreateUser method
-		token, err := userService.CreateUser("testuser", "password123")
+		err := userService.CreateUser("testuser", "password123")
 
 		// Assertions
 		assert.Error(t, err)
-		// Token should be empty
-		assert.Empty(t, token)
 	})
+
+	t.Run("Should return error if password longer than 72 characters", func(t *testing.T) {
+		// Call the CreateUser method
+		err := userService.CreateUser("testuser", "arandompasswordthatislongerthan72characterslongarandompasswordthatislongerthan72characterslongarandompasswordthatislongerthan72characterslong")
+
+		// Assertions
+		assert.Error(t, err)
+	})
+
 }
 
 func TestLoginUser(t *testing.T) {
@@ -41,32 +47,28 @@ func TestLoginUser(t *testing.T) {
 	// Create a new instance of AuthService with the mock repository
 	userService := NewAuthService(mockUserRepository)
 
-	user := &models.User{Username: "test", Password: "password123"}
+	user := &user_model.User{Username: "test", Password: "password123"}
 
 	t.Run("Login User Successfully", func(t *testing.T) {
-		_, err := userService.CreateUser(user.Username, user.Password)
+		err := userService.CreateUser(user.Username, user.Password)
 		assert.NoError(t, err)
 
-		// Should login successfully
-		token, err := userService.LoginUser(user.Username, user.Password)
+		// Should log in successfully
+		err = userService.LoginUser(user.Username, user.Password)
 		assert.NoError(t, err)
-		assert.NotEmpty(t, token)
 	})
 
 	t.Run("Failed to Login User", func(t *testing.T) {
-		// Should fail to login
-		token, err := userService.LoginUser(user.Username, "wrongpassword")
+		// Should fail to log in
+		err := userService.LoginUser(user.Username, "wrongpassword")
 		assert.Error(t, err)
-		assert.Empty(t, token)
-
 	})
 
 	t.Run("User not found", func(t *testing.T) {
 
-		// Should fail to login
-		token, err := userService.LoginUser("unknown", "password")
+		// Should fail to log in
+		err := userService.LoginUser("unknown", "password")
 		assert.Error(t, err)
-		assert.Empty(t, token)
 
 	})