+++ CHANGELOG.md
@@ -1,6 +1,54 @@
 # Changelog
 All significant updates to this project will be meticulously documented in this log.
 
+## 0.3.0 - 09/03/2024
+
+### Added
+
+- **URL Repository:** Added a URL repository to the project.
+
+- **URL Shorten API:** Added a new route to the echo server to handle URL shortening.
+
+## 0.2.0.alpha.2 - 09/03/2024
+
+### Added
+
+- **URL Service:** Added a URL service to the project.
+
+- **Generate Short URL:** Added a function to generate a short URL.
+
+- **Mock URL Repository:** Added a mock URL repository to the project.
+
+## 0.2.0.alpha.1 - 09/03/2024
+
+### Added
+
+- **URL Handler:** Added a URL handler to the project.
+
+- **URL Model:** Added a URL model to the project.
+
+## 0.2.0 - 09/03/2024
+
+### Added
+
+- **API Documentation:** Added API documentation to the project.
+
+### Changed
+
+- **Directory of User Model:** Moved the user model to the `internal/app/models/user` directory.
+
+- **Directory of Auth Handler:** Moved the auth handler to the `internal/app/handler/auth` directory.
+
+- **Directory of Auth Repository:** Moved the auth repository to the `internal/app/repository/auth` directory.
+
+- **Directory of Auth Service:** Moved the auth service to the `internal/app/services/auth` directory.
+
+- **Make Token as a Service:** Refactored the code to make token generation a service.
+
+- **Handler Generates Token**: Refactored the code to make the handler generate tokens instead of auth service.
+
+- **Auth Handler Takes Token Service**: Refactored the code to make the auth handler take the token service as a dependency.
+
 ## 0.1.0 - 08/03/2024
 
 ### Changed